requirements-render.txt
# Production requirements for Render deployment
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
google-generativeai==0.8.5
pydantic==2.11.7
python-multipart==0.0.20
python-dotenv==1.1.1

# Celery and task queue dependencies
celery==5.3.1
redis==6.4.0
kombu==5.5.4
billiard==4.2.1
vine==5.1.0

# Additional production dependencies
gunicorn==21.2.0

# Core dependencies that might be needed
httpx==0.28.1
websockets==15.0.1

# Additional dependencies from your full requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.2
packaging==25.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==5.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
